requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coverage==5.5
cryptography==3.4.7
Django==3.2.4
django-countries==7.2.1
django-filter==2.4.0
django-mirage-field==1.1.8
django-otp==1.0.6
django-phonenumber-field==5.2.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-core==1.7.2
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.3
googleapis-common-protos==1.53.0
idna==3.2
model-bakery==1.3.2
packaging==21.0
phonenumbers==8.12.26
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyotp==2.6.0
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
testcase==0.1.0
tqdm==4.62.0
urllib3==1.26.6
